init-ifupdown: do things in a more generic way
This should be reusable for all boards.

+++ recipes-core/init-ifupdown/init-ifupdown_%.bbappend
@@ -1,7 +1,9 @@
 FILESEXTRAPATHS_prepend := "${THISDIR}:"
 
-SRC_URI += "file://can-pre-up"
+SRC_URI_ti33x += "file://can-pre-up"
 
-do_install_append() {
-	install -m 0755 ${WORKDIR}/can-pre-up ${D}${sysconfdir}/network/can-pre-up
+do_install_append () {
+    if [ -e ${WORKDIR}/can-pre-up ] ; then
+        install -m 0755 ${WORKDIR}/can-pre-up ${D}${sysconfdir}/network/can-pre-up
+    fi
 }
